+++ src/lib/project.js
@@ -2,7 +2,6 @@ import chalk from 'chalk';
 import { spawnSync } from 'child_process';
 import enquirer from 'enquirer';
 import fs from 'fs-extra';
-import gittar from 'gittar';
 import ora from 'ora';
 import path from 'path';
 import shell from 'shelljs';
@@ -14,6 +13,7 @@ import nuxtGradientBackground from '../lib/ui/nuxt/nuxtGradientBackground.js';
 import customLayoutSvelte from '../lib/ui/svelte/customLayoutSvelte.js';
 import customPageSvelte from '../lib/ui/svelte/customPageSvelte.js';
 import Constants from './constants.js';
+import { setupProject } from './helpers.js';
 import gradientBackground from './ui/svelte/gradientBackground.js';
 
 const __filename = url.fileURLToPath(import.meta.url);
@@ -40,6 +40,7 @@ export async function project({ name, ui }) {
     shell.exit(1);
   }
 
+  const lang = 'ts';
   let res;
   if (!ui) {
     try {
@@ -121,7 +122,9 @@ export async function project({ name, ui }) {
     shell.mkdir('contracts');
     shell.cd('contracts');
   }
-  if (!(await fetchProjectTemplate())) shell.exit(1);
+  if (!(await setupProject(shell.pwd().toString(), lang))) {
+    shell.exit(1);
+  }
 
   await step(
     'NPM install',
@@ -149,50 +152,6 @@ export async function project({ name, ui }) {
 
   console.log(chalk.green(str));
   process.exit(0);
-}
-
-/**
- * Fetch project template.
- * @returns {Promise<boolean>} - True if successful; false if not.
- */
-async function fetchProjectTemplate() {
-  const projectName = 'project-ts';
-
-  const step = 'Set up project';
-  const spin = ora({ text: `${step}...`, discardStdin: true }).start();
-
-  try {
-    const src = 'github:o1-labs/zkapp-cli#main';
-    await gittar.fetch(src, { force: true });
-
-    // Note: Extract will overwrite any existing dir's contents. Ensure
-    // destination does not exist before this.
-    const TEMP = '.gittar-temp-dir';
-    await gittar.extract(src, TEMP, {
-      filter(path) {
-        return path.includes(`templates/${projectName}/`);
-      },
-    });
-
-    // Copy files into current working dir
-    shell.cp(
-      '-r',
-      `${path.join(TEMP, 'templates', projectName)}${path.sep}.`,
-      '.'
-    );
-    shell.rm('-r', TEMP);
-
-    // Create a keys dir because it's not part of the project scaffolding given
-    // we have `keys` in our .gitignore.
-    shell.mkdir('keys');
-
-    spin.succeed(chalk.green(step));
-    return true;
-  } catch (err) {
-    spin.fail(step);
-    console.error(err);
-    return false;
-  }
 }
 
 /**